requirement.txt
keras==3.10.0
tensorflow==2.19.0
numpy==2.0.2
pandas==2.2.2
matplotlib==3.10.0
seaborn==0.13.2
Pillow==11.3.0
IPython==7.34.0
